+++ src/utilities.ts
@@ -1,3 +1,5 @@
+import type { IgnoredUpdater } from '@vueuse/core'
+import type { Ref, WatchStopHandle } from 'vue'
 import type {
     ParamMap,
     StoreRepositoryReadOptions,
@@ -5,7 +7,7 @@ import type {
 } from './types'
 import {
     debounceFilter,
-    type IgnoredUpdater,
+
     useDocumentVisibility,
     useWindowFocus,
     watchIgnorable,
@@ -13,12 +15,11 @@ import {
 import {
     computed,
     isRef,
-    type Ref,
     ref,
     unref,
     watch,
     watchEffect,
-    type WatchStopHandle,
+
 } from 'vue'
 import { StoreRepositoryStatus } from './constants'